mediacapture-image/MediaStreamTrack-getConstraints-fast.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://w3c.github.io/mediacapture-main/#dom-mediastreamtrack-applyconstraints">
<link rel="help" href="https://w3c.github.io/mediacapture-image/#extensions">
<body>
<canvas id='canvas' width=10 height=10/>
</body>
<script>

const constraints = { whiteBalanceMode     : "manual",
                      exposureMode         : "manual",
                      focusMode            : "single-shot",

                      exposureCompensation : 133.77,
                      exposureTime         : 10000, // in nano-seconds.
                      colorTemperature     : 6000,
                      iso                  : 120.0,

                      brightness           : 3,
                      contrast             : 4,
                      saturation           : 5,
                      sharpness            : 6,
                      focusDistance        : 7,

                      pan                  : 8,
                      tilt                 : 9,

                      zoom                 : 3.141592
                      // TODO: torch https://crbug.com/700607.
                    };

var canvas = document.getElementById('canvas');
var context = canvas.getContext("2d");
context.fillStyle = "red";
context.fillRect(0, 0, 10, 10);

// These tests verify that MediaStreamTrack.getConstraints() exists and that,
// returns the constraints passed beforehand with applyConstraints.
function makePromiseTest(name, c) {
  promise_test(async function() {
    var stream = canvas.captureStream();
    var videoTrack = stream.getVideoTracks()[0];

    const constraintsIn = { advanced : [ c ]};

    try {
      await videoTrack.applyConstraints(constraintsIn);
    } catch(e) {
      // If applyConstraints() rejects, continue and assume that |constraintsIn|
      // are stored anyway. Note that there doesn't appear to be any mention of
      // this beahvior in the specs. TODO: remove this try/catch block.
    }
    const constraintsOut = videoTrack.getConstraints();
    assert_object_equals(constraintsOut, constraintsIn, "constraints");

    // Clear constraints by sending an empty constraint set.
    await videoTrack.applyConstraints({});
    const constraintsOut2 = videoTrack.getConstraints();
    assert_object_equals(constraintsOut2, {}, "constraints");
  }, name);
}

// Send each line of |constraints| in turn and then the whole dictionary.
for (key in constraints) {
  var one_constraint = {};
  one_constraint[key] = constraints[key];
  makePromiseTest('MediaStreamTrack.getConstraints(), key: ' + key, one_constraint);
}

makePromiseTest('MediaStreamTrack.getConstraints(), complete ', constraints);

</script>
